use smaller canvas
